+++ lib/pgvector.dart
@@ -2,9 +2,11 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'halfvec.dart';
+import 'sparsevec.dart';
 import 'vector.dart';
 
 export 'halfvec.dart' show HalfVector;
+export 'sparsevec.dart' show SparseVector;
 export 'vector.dart' show Vector;
 
 class Pgvector {