+++ test/postgres_test.dart
@@ -17,18 +17,21 @@ void main() {
     await connection.execute("DROP TABLE IF EXISTS items");
 
     await connection.execute(
-        "CREATE TABLE items (id bigserial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3))");
+        "CREATE TABLE items (id bigserial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3), sparse_embedding sparsevec(3))");
 
     await connection.execute(
         Sql.named(
-            "INSERT INTO items (embedding, half_embedding) VALUES (@a, @d), (@b, @e), (@c, @f)"),
+            "INSERT INTO items (embedding, half_embedding, sparse_embedding) VALUES (@a, @d, @g), (@b, @e, @h), (@c, @f, @i)"),
         parameters: {
           "a": Vector([1, 1, 1]).toString(),
           "b": Vector([2, 2, 2]).toString(),
           "c": Vector([1, 1, 2]).toString(),
           "d": HalfVector([1, 1, 1]).toString(),
           "e": HalfVector([2, 2, 2]).toString(),
-          "f": HalfVector([1, 1, 2]).toString()
+          "f": HalfVector([1, 1, 2]).toString(),
+          "g": SparseVector([1, 1, 1]).toString(),
+          "h": SparseVector([2, 2, 2]).toString(),
+          "i": SparseVector([1, 1, 2]).toString()
         });
 
     List<List<dynamic>> results = await connection.execute(